Improve price list
- handle missing image
- refresh after adding a product

+++ app/src/main/java/io/lab10/vallet/models/Product.kt
@@ -7,7 +7,7 @@ import io.objectbox.annotation.Entity
 import io.objectbox.annotation.Id
 
 @Entity
-data class Product(@Id var id: Long, var name: String, var price: Int, var imagePath: String, var localImagePath: String, var nfcTagId: String)  {
+data class Product(@Id var id: Long, var name: String, var price: Long, var imagePath: String, var localImagePath: String, var nfcTagId: String)  {
 
     override fun toString(): String {
         return name
